Fixed undefined functions in new tools.js
